Extract patchmods to individual files.
